+++ test/integration/helpers/model-expect.ts
@@ -179,7 +179,7 @@ export interface ExpectedEventBasedGatewayModelElement extends ExpectedShapeMode
   gatewayKind?: ShapeBpmnEventBasedGatewayKind;
 }
 
-export const bpmnVisualization = new BpmnVisualization({ container: null });
+export const bpmnVisualization = new BpmnVisualization(null);
 
 export function getDefaultParentId(): string {
   return bpmnVisualization.graph.getDefaultParent().id;
